% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_drat.R
\name{make_drat_bin_placeholders}
\alias{make_drat_bin_placeholders}
\title{Make binary placeholders for a drat repository}
\usage{
make_drat_bin_placeholders()
}
\description{
The goal of drat is to let you create a source repository. Supporting
the inclusion of binary packages is beyond the scope of these tools.
However, if you do not have the directory structure for binaries,
you will get some annoying warnings each time you install packages.
This function will help you get rid of those messages.
}
\examples{
  make_drat_bin_placeholders()
}
