% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_github_enterprise.R
\name{use_install_ghe}
\alias{use_install_ghe}
\alias{use_ghe_pat}
\alias{use_use_ghe}
\title{Use GitHub Enterprise components in your package}
\usage{
use_install_ghe(host, suffix, name)

use_ghe_pat(suffix, name)

use_use_ghe(host, suffix, name)
}
\arguments{
\item{host}{\code{character} hostname for GHE instance}

\item{suffix}{\code{character} lowercase suffix to use for function names}

\item{name}{\code{character} title-case name of the GHE instance}
}
\value{
\code{invisible(NULL)}, called for side effects
}
\description{
Use GitHub Enterprise components in your package
}
