% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_drat.R
\name{make_drat_bin_placeholders}
\alias{make_drat_bin_placeholders}
\title{Make binary placeholders for a drat repository}
\usage{
make_drat_bin_placeholders()
}
\description{
The goal of drat is to let you create a source repository. Supporting
the inclusion of binary packages is beyond the scope of these tools.
However, if your drat repository does not have the directory structure
for binaries, you will get some annoying warnings each time
you install packages. This function can help you get rid of those messages.
}
\details{
In a CRAN-like repository, binary packages are kept in their own directory
structure. This function looks in your project to see what parts of this
directory structure exist. It creates those parts of the structure that
do not already exist; but nothing is overwritten.
}
\examples{
\dontrun{
  make_drat_bin_placeholders()
}
}
